WIP continuing building up the alignment plan stack
