+++ startup/scripts/scratch.py
@@ -1726,71 +1726,6 @@ def run_alignment_scans_for_crystal_bundle(crystal=None, alignment_by=None, pil1
 
 
 
-def align_johann_spectrometer_plan_bundle(alignment_by=None, pil100k_roi_num=None,
-                                          alignment_data=None,
-                                          scan_range_roll=None, scan_range_yaw=None, velocity=None,
-                                          yaw_shift=None,
-                                          tweak_motor_range=None, tweak_motor_num_steps=None,
-                                          plot_func=None, liveplot_kwargs=None):
-    plans = []
-
-    if alignment_data is None:
-        alignment_data = johann_emission.alignment_data
-
-    enabled_crystals = [_c for _c, _e in johann_emission.enabled_crystals.items() if _e]
-
-    # enforce that the main crystal is aligned first
-    if 'main' in enabled_crystals:
-        enabled_crystals.pop(enabled_crystals.index('main'))
-        enabled_crystals = ['main'] + enabled_crystals
-
-    for crystal in enabled_crystals:
-        if crystal not in alignment_data.keys():
-            alignment_data[crystal] = []
-
-        crystals_not_being_aligned = [c for c in enabled_crystals if c != crystal]
-
-        plans.append({'plan_name': 'print_message_plan',
-                      'plan_kwargs': {
-                          'msg': f'Aligning {crystal} crystal. Moving other crystals from the field of view.',
-                          'add_timestamp': True,
-                          'tag': 'Spectrometer'}})
-        for cr in crystals_not_being_aligned:
-            yaw_direction = 1 if cr in ['aux3', 'aux5'] else -1
-            plans.append({'plan_name': 'move_relative_motor_plan',
-                          'plan_kwargs': {'motor_attr': _crystal_alignment_dict[cr]['yaw'],
-                                          'based_on': 'description',
-                                          'position': yaw_shift * yaw_direction}})
-        multiplier = 1 if crystal == 'main' else 1000
-        plans.append({'plan_name': 'run_alignment_scans_for_crystal_bundle',
-                      'plan_kwargs': {'crystal': crystal,
-                                      'alignment_by': alignment_by,
-                                      'pil100k_roi_num': pil100k_roi_num,
-                                      'alignment_data': alignment_data[crystal],
-                                      'scan_range_roll': scan_range_roll,
-                                      'scan_range_yaw': scan_range_yaw,
-                                      'velocity': velocity,
-                                      'tweak_motor_range': tweak_motor_range * multiplier,
-                                      'tweak_motor_num_steps': tweak_motor_num_steps,
-                                      'plot_func': plot_func,
-                                      'liveplot_kwargs': liveplot_kwargs}})
-
-        for cr in crystals_not_being_aligned:
-            yaw_direction = -1 if cr in ['aux3', 'aux5'] else 1
-            plans.append({'plan_name': 'move_relative_motor_plan',
-                          'plan_kwargs': {'motor_attr': _crystal_alignment_dict[cr]['yaw'],
-                                          'based_on': 'description',
-                                          'position': yaw_shift * yaw_direction}})
-
-        plans.append({'plan_name': 'print_message_plan',
-                      'plan_kwargs': {
-                          'msg': f'Done with aligning {crystal} crystal. Moving other crystals back into the field of view.',
-                          'add_timestamp': True,
-                          'tag': 'Spectrometer'}})
-
-
-
-    return plans
